# Make TopKQueue thread-safe. Get rid of similarityHolder.