This PR makes a `TopKQueue` thread-safe and reuses it in different collectors. There are some improvement diffs:

```
benchmark                              old ns/op     new ns/op     delta
BenchmarkSuggest-8                     17364         15893         -8.47%
BenchmarkAutoComplete-8                12175         12366         +1.57%
BenchmarkRealExampleInMemory-8         42058         41371         -1.63%
BenchmarkRealExampleOnDisc-8           42134         41990         -0.34%
BenchmarkSuggestWordsOnDisc-8          115070        102630        -10.81%
BenchmarkAutocompleteWordsOnDisc-8     48066         47897         -0.35%

benchmark                              old allocs     new allocs     delta
BenchmarkRealExampleInMemory-8         331            267            -19.34%
BenchmarkRealExampleOnDisc-8           331            267            -19.34%
BenchmarkSuggestWordsOnDisc-8          604            547            -9.44%
BenchmarkAutocompleteWordsOnDisc-8     315            315            +0.00%

benchmark                              old bytes     new bytes     delta
BenchmarkRealExampleInMemory-8         15526         13476         -13.20%
BenchmarkRealExampleOnDisc-8           15543         13500         -13.14%
BenchmarkSuggestWordsOnDisc-8          21933         20067         -8.51%
BenchmarkAutocompleteWordsOnDisc-8     12666         12699         +0.26%
```